resolve some conflict
