+++ src/main/java/SubCommands/Build.java
@@ -1,4 +1,4 @@
-package java.SubCommands;
+package SubCommands;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -7,15 +7,26 @@ import java.nio.file.StandardOpenOption;
 import java.util.Map;
 import java.util.concurrent.Callable;
 
+import FileConvertor.ConvertorForYaml;
+import FileConvertor.ConvertorToHtml;
+import Parser.PageParser;
 import Watcher.Watcher;
 import com.github.jknack.handlebars.Template;
-import java.FileConvertor.ConvertorForYaml;
-import java.FileConvertor.ConvertorToHtml;
-import java.Parser.PageParser;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 
 
+/**
+ * Build a static site from a specified path. Once the build done, all
+ * .md and .yaml files have been converted in a .html file in a new
+ * directory named build. Other files still unchanged.
+ *
+ * @author Dimitri De Bleser
+ * @author André Marques Nora
+ * @author Vincent Peer
+ * @author Ivan Vecerina
+ * @version 1.0
+ */
 @Command(name = "build", description = "Build a static site")
 public class Build implements Callable<Integer> {
 
@@ -25,27 +36,33 @@ public class Build implements Callable<Integer> {
     @CommandLine.Option(names = {"-w", "--watch"}, description = "build site for every update")
     private static boolean beingWatched;
 
+    /**
+     * Method to carry out the fonctionalities of build command
+     * @return Ok status if build terminated with success
+     * @throws IOException
+     */
     @Override public Integer call() throws IOException {
 
+        // start if the user use the option watch
         if(beingWatched){
             var watcher = new Watcher(site);
-            watcher.watch(String.valueOf(Path.of(site.toString())));
+            watcher.watch(String.valueOf(Path.of(site.toString()))); //watch the directory for changes
             beingWatched = false;
         }
 
         Map<String, Object> configuration = ConvertorForYaml.parseYaml(site);
-
         Template template = ConvertorToHtml.getMdTemplate(site);
 
+        new CommandLine(new Clean()).execute(site.toString());
 
-        new CommandLine(new Clean()).execute(site.toString()+"/build");
-
+        // browse the directory
         Files.walk(site)
                 .filter(file -> file.toString().endsWith(".md"))
                 .forEach(source -> {
                     try {
                         String html = PageParser.parse(site, configuration, template);
 
+                        //create html file when there is a md file
                         Path target = site.resolve("build")
                                 .resolve(site.relativize(source).toString().replace(".md", ".html"));
                         Files.createDirectories(target.getParent());
@@ -60,4 +77,3 @@ public class Build implements Callable<Integer> {
         return CommandLine.ExitCode.OK;
     }
 }
-
