+++ src/test/java/SubCommands/CleanTest.java
@@ -1,7 +1,7 @@
 package SubCommands;
 
 
-import java.Statique.Statique;
+import java.Statique;
 import java.SubCommands.Utils;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
